Move IO util to new file and implement copy WAL func
